backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
python-dotenv==1.0.0
bcrypt==4.1.1
requests==2.31.0
Werkzeug==3.0.1
SQLAlchemy==2.0.48